fix desktop tests
