+++ .circleci/config.yml
@@ -48,7 +48,7 @@ jobs:
   test:
     docker:
       - image: circleci/node:10.13
-    parallelism: 3
+    parallelism: 2
     steps:
       - yarn/setup
       - run:
