+++ src/desktop/apps/article/templates/__tests__/meta.jest.js
@@ -4,6 +4,7 @@ import jade from "jade"
 import fs from "fs"
 import Article from "desktop/models/article.coffee"
 import fixtures from "desktop/test/helpers/fixtures.coffee"
+import { StandardArticle } from "@artsy/reaction/dist/Components/Publishing/Fixtures/Articles"
 
 const render = templateName => {
   const filename = path.resolve(__dirname, `../${templateName}.jade`)
@@ -12,7 +13,7 @@ const render = templateName => {
 
 describe("Meta template", () => {
   it("contains basic meta tags", () => {
-    const article = _.extend({}, fixtures.article, {
+    const article = _.extend({}, StandardArticle, {
       slug: "artsy-editorial-slug",
       indexable: true,
       description: "Artsy Editorial is an editorial channel.",
@@ -27,26 +28,24 @@ describe("Meta template", () => {
         CURRENT_PATH: "/article/artsy-editorial-slug",
       },
     })
-    html.should.containEql(
-      "<title>Top Ten Booths at miart 2014 - Artsy</title>"
-    )
-    html.should.containEql(
+    expect(html).toMatch("<title>New York's Next Art District - Artsy</title>")
+    expect(html).toMatch(
       '<link rel="canonical" href="https://artsy.net/article/artsy-editorial-slug"/>'
     )
-    html.should.containEql(
+    expect(html).toMatch(
       '<meta name="description" content="Artsy Editorial is an editorial channel."/>'
     )
-    html.should.containEql(
+    expect(html).toMatch(
       '<meta property="og:description" content="Artsy Editorial is socially friendly."/>'
     )
-    html.should.containEql(
+    expect(html).toMatch(
       '<meta property="og:image" content="http://kitten-social.com"/>'
     )
-    html.should.not.containEql('<meta name="robots" content="noindex"/>')
+    expect(html).not.toMatch('<meta name="robots" content="noindex"/>')
   })
 
   it("uses a different title extension for news", () => {
-    const article = _.extend({}, fixtures.article, {
+    const article = _.extend({}, StandardArticle, {
       slug: "artsy-editorial-slug",
       description: "Artsy Editorial is an editorial channel.",
       social_description: "Artsy Editorial is socially friendly.",
@@ -61,10 +60,11 @@ describe("Meta template", () => {
         CURRENT_PATH: "/news/artsy-editorial-slug",
       },
     })
-    html.should.containEql(
-      "<title>Top Ten Booths at miart 2014 - Artsy News</title>"
+
+    expect(html).toMatch(
+      "<title>New York's Next Art District - Artsy News</title>"
     )
-    html.should.containEql(
+    expect(html).toMatch(
       '<link rel="canonical" href="https://artsy.net/news/artsy-editorial-slug"/>'
     )
   })
@@ -72,7 +72,7 @@ describe("Meta template", () => {
 
 describe("Classic meta template", () => {
   it("contains basic meta tags", () => {
-    const article = _.extend({}, fixtures.article, {
+    const article = _.extend({}, StandardArticle, {
       slug: "gallery-insights-slug",
       indexable: true,
       description: "Gallery Insights is a team channel.",
@@ -87,19 +87,20 @@ describe("Classic meta template", () => {
         APP_URL: "https://artsy.net",
       },
     })
-    html.should.containEql("<title>Top Ten Booths at miart 2014</title>")
-    html.should.containEql(
+
+    expect(html).toMatch("<title>New York's Next Art District</title>")
+    expect(html).toMatch(
       '<link rel="canonical" href="http://artsy.net/article/gallery-insights-slug"/>'
     )
-    html.should.containEql(
+    expect(html).toMatch(
       '<meta property="description" content="Gallery Insights is a team channel."/>'
     )
-    html.should.containEql(
+    expect(html).toMatch(
       '<meta name="og:description" content="Gallery Insights is a socially friendly team channel."/>'
     )
-    html.should.containEql(
+    expect(html).toMatch(
       '<meta property="og:image" content="http://kitten-social.com"/>'
     )
-    html.should.not.containEql('<meta name="robots" content="noindex"/>')
+    expect(html).not.toMatch('<meta name="robots" content="noindex"/>')
   })
 })
