+++ src/desktop/apps/partner2/test/client/shows_grid.test.coffee
@@ -121,8 +121,8 @@ describe 'PartnerShowsGridView', ->
           params: {}
         }, =>
           PartnerShowsGridView = mod = benv.requireWithJadeify(
-              (resolve __dirname, '../../client/shows_grid'), ['template', 'showFiguresTemplate']
-            )
+            (resolve __dirname, '../../client/shows_grid'), ['template', 'showFiguresTemplate']
+          )
           @src = [
             fabricate('show', { name: 'show1' } ),
             fabricate('show', { name: 'show2', featured: true } ),
