+++ src/desktop/test/models/feature.test.coffee
@@ -61,11 +61,11 @@ describe 'Feature', ->
 
     it 'fetches until end for sets whose items are featured links', (done) ->
       @feature.fetchSets success: (sets) ->
-          sets[0].get('type').should.equal 'featured links'
-          sets[0].get('name').should.equal 'Explore this bidness top'
-          sets[0].get('data').first().get('title').should.equal 'Featured link for this awesome page'
-          sets[0].get('data').should.have.lengthOf 12
-          done()
+        sets[0].get('type').should.equal 'featured links'
+        sets[0].get('name').should.equal 'Explore this bidness top'
+        sets[0].get('data').first().get('title').should.equal 'Featured link for this awesome page'
+        sets[0].get('data').should.have.lengthOf 12
+        done()
 
       _.last(Backbone.sync.args)[2].success([fabricate('set', name: 'Explore this bidness top', key: '0hello', id: 'def') ])
