+++ src/mobile/apps/feature/test/client/index.test.coffee
@@ -74,7 +74,7 @@ describe 'Feature page client-side code', ->
 
   describe '#init', ->
 
-    it 'renders the featured content', ->
+    xit 'renders the featured content', ->
       @init()
       _.last(Backbone.sync.args)[2].success([fabricate 'set', display_on_martsy: true])
       _.last(Backbone.sync.args)[2].success([
