+++ src/test/lib/middleware/proxy_to_gravity.test.coffee
@@ -13,8 +13,6 @@ gravity.get '/post', (req, res) ->
 startServer = (callback) ->
   envVars =
     NODE_ENV: 'test'
-    API_URL: "http://localhost:5001"
-    API_URL: "http://localhost:5001"
     APP_URL: "http://localhost:5000"
     PORT: 5000
   envVars[k] = val for k, val of process.env when not envVars[k]?
@@ -23,7 +21,7 @@ startServer = (callback) ->
     stdio: ["ipc"]
     env: envVars
   child.on "message", callback
-closeServer = => child.kill()
+closeServer = -> child.kill()
 
 xdescribe 'Setup', ->
